# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Conflict & Duplication Detector API",
    "description": "HTTP API for document ingestion, knowledge-base analysis, file checks, and chat.\n\nLong-running operations (ingest, analyse, check, chat) return **202 Accepted** with a job ID.\nPoll **GET /api/jobs/{jobId}** until status is `completed` or `failed`.",
    "version": "v1"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "Detector"
        ],
        "summary": "Health check",
        "description": "Returns API health and knowledge-base summary.",
        "operationId": "GetHealth",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/knowledge-base": {
      "get": {
        "tags": [
          "Detector"
        ],
        "summary": "Knowledge base status",
        "description": "Returns whether the vector store exists and how many chunks are loaded.",
        "operationId": "GetKnowledgeBase",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/documents": {
      "post": {
        "tags": [
          "Detector"
        ],
        "summary": "Ingest a document",
        "description": "Upload a document and enqueue ingestion into the knowledge base. Returns 202 with a job ID.",
        "operationId": "IngestDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/analysis": {
      "post": {
        "tags": [
          "Detector"
        ],
        "summary": "Analyse the knowledge base",
        "description": "Enqueue analysis of all ingested documents (duplications, conflicts, inconsistencies). Returns 202 with a job ID.",
        "operationId": "StartAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/check": {
      "post": {
        "tags": [
          "Detector"
        ],
        "summary": "Check a file against the knowledge base",
        "description": "Upload a file and compare it against the ingested knowledge base. Returns 202 with a job ID.",
        "operationId": "CheckDocument",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": [
          "Detector"
        ],
        "summary": "Chat with the knowledge base",
        "description": "Enqueue a natural-language query against ingested documents. Returns 202 with a job ID.",
        "operationId": "StartChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "Detector"
        ],
        "summary": "List recent jobs",
        "operationId": "ListJobs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/jobs/{jobId}": {
      "get": {
        "tags": [
          "Detector"
        ],
        "summary": "Get job status and result",
        "description": "Poll this endpoint until status is completed or failed.",
        "operationId": "GetJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisJobRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "topic": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request body for knowledge-base analysis."
      },
      "ChatJobRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request body for chat against the knowledge base."
      }
    }
  }
}